+++ diagnostics/ito_diagonal.py
@@ -16,7 +16,7 @@ import os
 
 import torch
 
-from tests.problems import Ex2
+from tests.problems import Ex1
 from torchsde import BrownianInterval
 from torchsde.settings import LEVY_AREA_APPROXIMATIONS
 from . import inspection
@@ -28,7 +28,7 @@ def main():
     t0, t1, steps, dt = 0., 2., 10, 1e-1
     ts = torch.linspace(t0, t1, steps=steps, device=device)
     dts = tuple(2 ** -i for i in range(1, 7))  # For checking strong order.
-    sde = Ex2(d=d).to(device)
+    sde = Ex1(d=d).to(device)
     methods = ('euler', 'srk', 'milstein', 'milstein')
     options = (None, None, None, dict(grad_free=True))
     labels = ('euler', 'srk', 'milstein', 'grad-free milstein')
