+++ torchsde/_core/adjoint.py
@@ -198,21 +198,15 @@ def sdeint_adjoint(sde: nn.Module,
 
 
 def _select_default_adjoint_method(sde: base_sde.ForwardSDE, adjoint_method: str) -> str:
-    sde_type, noise_type = sde.sde_type, sde.noise_type
-
-    if adjoint_method is None:  # Select the default based on noise type of forward.
+    """Select the default method for adjoint computation based on the noise type of the forward SDE."""
+    if adjoint_method is None:
         adjoint_method = {
             SDE_TYPES.ito: {
                 NOISE_TYPES.diagonal: METHODS.milstein,
                 NOISE_TYPES.additive: METHODS.euler,
                 NOISE_TYPES.scalar: METHODS.euler,
-            }.get(noise_type, "unsupported"),
-            SDE_TYPES.stratonovich: {
-                NOISE_TYPES.general: METHODS.midpoint,
-            }.get(noise_type, "unsupported")
-        }[sde_type]
-
-        if adjoint_method == "unsupported":
-            raise ValueError(f"Adjoint not supported for {sde_type} SDEs with noise type {noise_type}.")
-
+                NOISE_TYPES.general: METHODS.euler,
+            }[sde.noise_type],
+            SDE_TYPES.stratonovich: METHODS.midpoint,
+        }[sde.sde_type]
     return adjoint_method
